Insertion Sort
First, the length of the given array is calculated. If the length of the array is 0 or 1, it is returned as it is since this type of array is already sorted and doesn't have enough elements to be compared to each other. 
For the arrays with more than one element, the function starts by iterating over the current element (the one at position 1), also called the Key, and compares it with the elements of the sorted list that precedes it. If an element in the 'sorted' portion is larger than the key, it is shifted to the right and a 'space' (which is actually the original element that was shifted or copied to the succeeding place) is created which is then occupied by the Key. This process continues till the Key finds it correct position. The iterator is used to find the next Key.